refactor(frontend): generalize filter controls
